Merged lifecycle object into testEnvironment
This makes the module declarations more concise but does not break the setup and teardown system. Compare:

module("old-style", {
  setup: function() { this.foo = "bar"; },
});

module("new-style", {
  foo:"bar"
});

In addition, you can now attach arbitrary meta-data to a module:

module("test metadata", {
	bugid: 5311,
	speed:"slow"
});

Motivation: I'm working on a CommonJS 'discovery' mode for qunit that first discovers all the tests in a suite, allows the user to filter and sort the tests by meta-data and then execute the selected tests.